Refactor: Move converter to model module
This is a preparation for the next commits
