+++ src/shogun2-core/shogun2-model/src/main/java/de/terrestris/shogun2/converter/PropertyValueConverter.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package de.terrestris.shogun2.util.converter;
+package de.terrestris.shogun2.converter;
 
 import javax.persistence.AttributeConverter;
