+++ src/shogun2-core/shogun2-model/src/test/java/de/terrestris/shogun2/converter/PropertyValueConverterTest.java
@@ -1,4 +1,4 @@
-package de.terrestris.shogun2.util.converter;
+package de.terrestris.shogun2.converter;
 
 import org.junit.Test;
 import static org.junit.Assert.*;
